requirements.txt
# Mobile Migration — FastAPI Backend
# Install:  pip install -r requirements.txt

fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<5.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
pandas>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-multipart>=0.0.6
python-dotenv>=1.0,<2.0
yfinance>=0.2,<1.0
APScheduler>=3.10,<4.0
slowapi>=0.1.9,<1.0
cryptography>=41.0,<44.0
gunicorn>=21.2,<23.0
google-generativeai>=0.8.0
PyMuPDF>=1.24.0
Pillow>=10.0
